+++ data/tasks.txt
@@ -1,2 +1,2 @@
-T | 1 | hello
-T | 0 | heyyy
+T | 0 | borrow book
+D | 0 | return book | 29/08/2019 1000
