--- src/main/java/AddCommand.java
@@ -1,2 +1,0 @@
-public abstract class AddCommand extends Command {
-}
