+++ src/main/java/duke/command/DeadlineCommand.java
@@ -1,3 +1,11 @@
+package duke.command;
+
+import duke.main.Storage;
+import duke.main.TaskList;
+import duke.main.UI;
+import duke.task.Deadline;
+import duke.task.Task;
+
 import java.util.Date;
 
 public class DeadlineCommand extends AddCommand {
