+++ src/main/java/duke/command/ExitCommand.java
@@ -1,3 +1,9 @@
+package duke.command;
+
+import duke.main.Storage;
+import duke.main.TaskList;
+import duke.main.UI;
+
 public class ExitCommand extends Command {
     @Override
     public boolean isExit() {
