+++ src/main/java/duke/command/TodoCommand.java
@@ -1,3 +1,11 @@
+package duke.command;
+
+import duke.main.Storage;
+import duke.main.TaskList;
+import duke.main.UI;
+import duke.task.Task;
+import duke.task.Todo;
+
 public class TodoCommand extends AddCommand {
     String description;
