+++ src/main/java/duke/error/DukeException.java
@@ -1,3 +1,5 @@
+package duke.error;
+
 public class DukeException extends Exception {
     public DukeException(String s) {
         super("\u2639 Sorry. " + s);
