+++ src/main/java/duke/error/FileLoadError.java
@@ -1,3 +1,5 @@
+package duke.error;
+
 public class FileLoadError extends DukeException {
     public FileLoadError() {
         super("There was a problem loading the saved list of tasks.");
