+++ src/main/java/duke/error/InsufficientArgumentError.java
@@ -1,3 +1,5 @@
+package duke.error;
+
 public class InsufficientArgumentError extends DukeException{
     public InsufficientArgumentError() {
         super("The description cannot be empty.");
