+++ src/main/java/duke/error/NoSuchTaskError.java
@@ -1,3 +1,5 @@
+package duke.error;
+
 public class NoSuchTaskError extends DukeException {
     public NoSuchTaskError() {
         super("There is no such task.");
