+++ src/main/java/duke/error/UnknownCommandError.java
@@ -1,3 +1,5 @@
+package duke.error;
+
 public class UnknownCommandError extends DukeException {
     public UnknownCommandError() {
         super("I don't know what that means.");
