+++ src/main/java/duke/main/Duke.java
@@ -1,0 +1,43 @@
+package duke.main;
+
+import duke.command.Command;
+import duke.error.DukeException;
+import duke.error.FileLoadError;
+import duke.error.IncorrectDateTimeError;
+
+public class Duke {
+    private Storage storage;
+    private TaskList tasks;
+    private UI ui;
+
+    public Duke(String filePath) {
+        ui = new UI();
+        storage = new Storage(filePath);
+        try {
+            tasks = new TaskList(storage.load());
+        } catch (FileLoadError | IncorrectDateTimeError e) {
+            ui.showError(e);
+            tasks = new TaskList();
+        }
+    }
+
+    public void run() {
+        ui.showWelcome();
+        boolean isExit = false;
+        while (!isExit) {
+            try {
+                String fullCommand = ui.readCommand();
+                Command c = Parser.parse(fullCommand);
+                c.execute(tasks, ui, storage);
+                isExit = c.isExit();
+            } catch (DukeException e) {
+                ui.showError(e);
+            }
+        }
+        ui.showBye();
+    }
+
+    public static void main(String[] args) {
+        new Duke("data/tasks.txt").run();
+    }
+}
